% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao.R
\name{encontrar_pior_ano_time}
\alias{encontrar_pior_ano_time}
\title{Encontrar pior ano do time}
\usage{
encontrar_pior_ano_time(time)
}
\arguments{
\item{time}{Nome do time}
}
\value{
Uma tibble
}
\description{
Essa função encontra o pior ano de um time
participante do Brasileirão.
}
\examples{
encontrar_pior_ano_time(time = "Flamengo")
}
