% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{partidas_brasileirao}
\alias{partidas_brasileirao}
\title{Jogos do Brasileirao}
\format{
Uma tibble com 8 colunas
\describe{
\item{temporada}{Descrição da coluna 1}
\item{data}{Descrição da coluna 2}
\item{time_casa}{Descrição da coluna 2}
...
}
}
\source{
Pacote Brasileirao \url{https://github.com/williamorim/brasileirao}
}
\usage{
partidas_brasileirao
}
\description{
Descrição da base de dados de jogos de futebol
do Campeonato Brasileirao
}
\keyword{datasets}
